requirements.txt
quickchart.io
langchain-upstage
streamlit